# api/openapi.yaml
# Code generated by swag. DO NOT EDIT.
# Source: make openapi (swag annotations on handlers).

basePath: /
definitions:
  chargepoint.ChargePoint:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      location:
        $ref: '#/definitions/chargepoint.Location'
      name:
        example: Dock 1
        type: string
      status:
        allOf:
        - $ref: '#/definitions/chargepoint.Status'
        enum:
        - AVAILABLE
        - OCCUPIED
        - OFFLINE
        example: AVAILABLE
    type: object
  chargepoint.CreateRequest:
    properties:
      location:
        $ref: '#/definitions/chargepoint.Location'
      name:
        example: Dock 1
        type: string
      status:
        allOf:
        - $ref: '#/definitions/chargepoint.Status'
        enum:
        - AVAILABLE
        - OCCUPIED
        - OFFLINE
        example: AVAILABLE
    type: object
  chargepoint.ErrorBody:
    properties:
      error:
        example: charge point not found
        type: string
    type: object
  chargepoint.Location:
    properties:
      lat:
        example: 52.37
        maximum: 90
        minimum: -90
        type: number
      lon:
        example: 4.89
        maximum: 180
        minimum: -180
        type: number
    type: object
  chargepoint.Status:
    enum:
    - AVAILABLE
    - OCCUPIED
    - OFFLINE
    type: string
    x-enum-varnames:
    - StatusAvailable
    - StatusOccupied
    - StatusOffline
host: localhost:8080
info:
  contact: {}
  description: Manage and query electric vehicle charge points.
  title: EV Charge Points API
  version: 1.0.0
paths:
  /chargepoints:
    post:
      consumes:
      - application/json
      description: Creates a charge point; server assigns a UUID (client id ignored)
      parameters:
      - description: Charge point
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/chargepoint.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chargepoint.ChargePoint'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
      summary: Create a charge point
      tags:
      - chargepoints
  /chargepoints/{id}:
    get:
      parameters:
      - description: Charge point UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chargepoint.ChargePoint'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
      summary: Get a charge point by ID
      tags:
      - chargepoints
  /chargepoints/nearby:
    get:
      description: Returns charge points within radius kilometers of lat/lon
      parameters:
      - description: Center latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Center longitude
        in: query
        name: lon
        required: true
        type: number
      - description: Search radius in kilometers
        in: query
        name: radius
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/chargepoint.ChargePoint'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
      summary: Find charge points within a radius
      tags:
      - chargepoints
  /healthz:
    get:
      description: Returns 200 when the process is up (does not check the database)
      produces:
      - text/plain
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Liveness probe
      tags:
      - ops
  /readyz:
    get:
      description: Returns 200 when the database is reachable
      produces:
      - text/plain
      responses:
        "200":
          description: ok
          schema:
            type: string
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/chargepoint.ErrorBody'
      summary: Readiness probe
      tags:
      - ops
swagger: "2.0"